DTV: use kHz for all frequencies
This is the simplest option. Hz resolution is useless for this purpose,
exhibits so many trailing zeroes that the preferences dialog overflows,
and overflows 32-bits values for satellite carrier frequencies.

This is also the default resolution for BDA. I see no need to copy the
"inconsistent" resolution if Linux-DVB (Hz for C/T, kHz for S).

For backward compatibility, frequencies beyond the radar bands are
divided by 1000. This corresponds to EHF and infrared in kHz.

+++ modules/access/dtv/dtv.h
@@ -45,9 +45,9 @@ int dvb_set_dvbc (dvb_device_t *, uint32_t freq, const char *mod,
                   uint32_t srate, const char *fec);
 
 /* DVB-S */
-int dvb_set_dvbs (dvb_device_t *, uint64_t freq, uint32_t srate,
+int dvb_set_dvbs (dvb_device_t *, uint32_t freq, uint32_t srate,
                   const char *fec);
-int dvb_set_dvbs2 (dvb_device_t *, uint64_t freq, const char *mod,
+int dvb_set_dvbs2 (dvb_device_t *, uint32_t freq, const char *mod,
                    uint32_t srate, const char *fec, int pilot, int rolloff);
 int dvb_set_sec (dvb_device_t *, bool tone, int voltage, bool high_voltage);
 /* XXX^^ */
